% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientFunctions.R
\docType{data}
\name{allMonths}
\alias{allMonths}
\title{The list of all months}
\format{
An object of class \code{character} of length 12.
}
\usage{
allMonths
}
\description{
The list of all months
}
\keyword{datasets}
