% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientConfiguration.R
\name{biosimclient.config}
\alias{biosimclient.config}
\title{Configure the client}
\usage{
biosimclient.config(
  forceClimateGenerationEnabled = NULL,
  nbNearestNeighbours = NULL,
  isTestModeEnabled = NULL
)
}
\arguments{
\item{forceClimateGenerationEnabled}{a logical}

\item{nbNearestNeighbours}{an integer}

\item{isTestModeEnabled}{a logical (only for test purpose)}
}
\description{
The forceClimateGenerationEnabled argument forces BioSIM to generate climate for past dates instead of
using the observations from the climate stations. By default this option is set to false. The
nbNearestNeighbours argument sets the number of stations for the imputation of climate variables.
}
\details{
If an argument is set to null, there is no effect at all. If all the arguments are set to null, then
the configuration is reset to its default value: the climate variables of past dates relies on observations
and the number of climate stations is set to 4.
}
\examples{
\dontrun{
### enables the climate generation for past dates and uses 20 climate stations
biosimclient.config(T, 20)

### reset the configuration
biosimclient.config() }

}
