% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientConfiguration.R
\name{biosimclient.getConfiguration}
\alias{biosimclient.getConfiguration}
\title{Report of the climate generation settings}
\usage{
biosimclient.getConfiguration()
}
\value{
a data.frame object
}
\description{
The isForceClimateGenerationEnabled setting forces BioSIM to generate climate for past dates
instead of using the observations from the climate stations. By default this option is set
to false. The nbNearestNeighbours setting is the number of stations used to impute climate variables to a
particular location.
}
\details{
All the settings can be changed through the biosimclient.config function.
}
