% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientConfiguration.R
\name{biosimclient.isForceClimateGenerationEnabled}
\alias{biosimclient.isForceClimateGenerationEnabled}
\title{Report of the climate generation option}
\usage{
biosimclient.isForceClimateGenerationEnabled()
}
\value{
a logical: true if the option is enabled or false otherwise
}
\description{
The climate generation option forces BioSIM to generate climate for past dates instead of
using the observations from the climate stations. By default this option is set to false.
The option can be set through the biosimclient.setForceClimateGenerationEnabled function.
}
