% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientConfiguration.R
\name{biosimclient.setForceClimateGenerationEnabled}
\alias{biosimclient.setForceClimateGenerationEnabled}
\title{Set the climate generation option}
\usage{
biosimclient.setForceClimateGenerationEnabled(bool)
}
\arguments{
\item{bool}{a logical}
}
\description{
The climate generation option forces BioSIM to generate climate for past dates instead of
using the observations from the climate stations. By default this option is set to false.
}
