% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientFunctions.R
\name{generateWeather}
\alias{generateWeather}
\title{Generate a meteorological time series and apply one or many models.}
\usage{
generateWeather(
  modelNames,
  fromYr,
  toYr,
  id,
  latDeg,
  longDeg,
  elevM = rep(NA, length(longDeg)),
  rep = 1,
  repModel = 1,
  rcp = "RCP45",
  climModel = "RCM4",
  additionalParms = NULL
)
}
\arguments{
\item{modelNames}{a character or a vector of character. Should be one or some models listed in the available models (see the getModelList() method)}

\item{fromYr}{the starting date (yr) of the period (inclusive)}

\item{toYr}{the ending date (yr) of the period (inclusive)}

\item{id}{a vector with the ids of the plots}

\item{latDeg}{the latitudes of the plots}

\item{longDeg}{the longitudes of the plots}

\item{elevM}{the elevations of the plots (can contain some NA or can be NULL, in which cases BioSim relies on a digital elevation model)}

\item{rep}{number of replicates of generated climate (is set to 1 by default)}

\item{repModel}{number of replicates on the model end (is set to 1 by default)}

\item{rcp}{an representative concentration pathway (either "CONSTANT_CLIMATE", "RCP45" or "RCP85")}

\item{climModel}{a climatic model (either "RCM4", "GCM4" or "Hadley")}

\item{additionalParms}{a list of named vectors with the additional parameters if needed}
}
\value{
a list of data.frame objects
}
\description{
This function generated a meteorological time series for some locations
and applies one or many models on this series. Default RCP and Climate
model are RCP45 and RCM4. The CONSTANT_CLIMATE rcp option actually
reproduces the climate observed between 1991 and 2020.
}
\examples{

locations <- BioSIM::twoLocationsInSouthernQuebec
addParms <- c("LowerThreshold"=5)
\dontrun{
degreeDays <- generateWeather("DegreeDay_Annual", 2017, 2021, locations$Name, locations$Latitude,
                   locations$Longitude, locations$Elevation,
                   rcp = "RCP85", climModel = "GCM4",
                   additionalParms = list(addParms))}

}
