% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientFunctions.R
\name{getAnnualNormals}
\alias{getAnnualNormals}
\title{Return the annual normals for a period}
\usage{
getAnnualNormals(period, variables, id, latDeg, longDeg, elevM)
}
\arguments{
\item{period}{a string representing the period (either "1951_1980", "1961_1990", "1971_2000" or "1981_2010")}

\item{variables}{the variables of interest typically a vector such as c("TN", "TX", "P") for minimum temperature, maximum temperature and precipitation}

\item{id}{a vector with the ids of the plots}

\item{latDeg}{the latitudes of the plots}

\item{longDeg}{the longitudes of the plots}

\item{elevM}{the elevations of the plots (can contain some NA, in which case BioSim relies on a digital elevation model)}
}
\value{
a data.frame object
}
\description{
Return the annual normals for a period
}
