% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientFunctions.R
\name{getClimateVariables}
\alias{getClimateVariables}
\title{Generates the climate for particular locations and applies a model on this generated climate.}
\usage{
getClimateVariables(
  fromYr,
  toYr,
  id,
  latDeg,
  longDeg,
  elevM,
  modelName,
  isEphemeral
)
}
\arguments{
\item{fromYr}{the starting date (yr) of the period (inclusive)}

\item{toYr}{the ending date (yr) of the period (inclusive)}

\item{id}{a vector with the ids of the plots}

\item{latDeg}{the latitudes of the plots}

\item{longDeg}{the longitudes of the plots}

\item{elevM}{the elevations of the plots (can contain some NA, in which case BioSim relies on a digital elevation model)}

\item{modelName}{a character. Should be one of the models listed in the available models (see the getModelList() method)}

\item{isEphemeral}{a logical. If set to true, the generated climate is not stored on the server, which implies a greater
computational burden and inconsistencies if different models are applied on the same locations.}
}
\description{
Generates the climate for particular locations and applies a model on this generated climate.
}
