% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientFunctions.R
\name{getModelDefaultParameters}
\alias{getModelDefaultParameters}
\title{Provide help for a particular model}
\usage{
getModelDefaultParameters(modelName)
}
\arguments{
\item{modelName}{should be one of the character string returned by the getModelList function}
}
\description{
Provide help for a particular model
}
\examples{

\dontrun{
getModelHelp("Spruce_Budworm_Biology")}

}
