% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientFunctions.R
\name{getModelOutput}
\alias{getModelOutput}
\title{Generates the climate for particular locations and applies a model on this generated climate.}
\usage{
getModelOutput(
  fromYr,
  toYr,
  id,
  latDeg,
  longDeg,
  elevM,
  modelName,
  isEphemeral,
  rep = 1,
  rcp = "RCP45",
  climModel = "RCM4",
  additionalParms = NULL
)
}
\arguments{
\item{fromYr}{the starting date (yr) of the period (inclusive)}

\item{toYr}{the ending date (yr) of the period (inclusive)}

\item{id}{a vector with the ids of the plots}

\item{latDeg}{the latitudes of the plots}

\item{longDeg}{the longitudes of the plots}

\item{elevM}{the elevations of the plots (can contain some NA, in which case BioSim relies on a digital elevation model)}

\item{modelName}{a character. Should be one of the models listed in the available models (see the getModelList() method)}

\item{isEphemeral}{a logical. If set to true, the generated climate is not stored on the server, which implies a greater
computational burden and inconsistencies if different models are applied on the same locations.}

\item{rep}{number of replicates of generated climate (is set to 1 by default)}

\item{rcp}{an representative concentration pathway (either "RCP45" or "RCP85")}

\item{climModel}{a climatic model (either "RCM4", "GCM4" or "Hadley")}

\item{additionalParms}{a named vector with the additional parameters if needed}
}
\value{
a data.frame object
}
\description{
Generates the climate for particular locations and applies a model on this generated climate.
}
\examples{

locations <- BioSIM::twoLocationsInSouthernQuebec
# degreeDays <- getClimateVariables(2017, 2021, locations$id, locations$latDeg,
#                                locations$longDeg, locations$elevM, "DegreeDay_Annual",
#                                F, rcp = "RCP85", climModel = "GCM4") ## not run


}
