% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientFunctions.R
\name{getNormals}
\alias{getNormals}
\title{Return the normals for a period}
\usage{
getNormals(
  period,
  variables,
  id,
  latDeg,
  longDeg,
  elevM,
  averageOverTheseMonths
)
}
\arguments{
\item{period}{a string representing the period (either "1951_1980", "1961_1990", "1971_2000" or "1981_2010")}

\item{variables}{the variables of interest typically a vector such as c("TN", "TX", "P") for minimum temperature, maximum temperature and precipitation}

\item{id}{a vector with the ids of the plots}

\item{latDeg}{the latitudes of the plots}

\item{longDeg}{the longitudes of the plots}

\item{elevM}{the elevations of the plots (can contain some NA, in which case BioSim relies on a digital elevation model)}

\item{averageOverTheseMonths}{a vector with some months if there is a need for aggregating the climate varibles}
}
\value{
a data.frame object
}
\description{
If the argument averageOverTheseMonths is left NULL or empty, the monthly normals are provided. If
this argument is filled with some months, then the normal are aggregated over these months.
}
