% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientConfiguration.R
\name{isMultithreadingEnabled}
\alias{isMultithreadingEnabled}
\title{Checks whether or not the multi threading is enabled.}
\usage{
isMultithreadingEnabled()
}
\value{
a logical
}
\description{
The multi threading allows to send multiple requests simultaneously. It
is handled by the back-end Java client. This method returns true if
this option is enabled (which is the default value).
}
