% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientFunctions.R
\name{setMultithreadingEnabled}
\alias{setMultithreadingEnabled}
\title{Enables or disables the multi threading.}
\usage{
setMultithreadingEnabled(enabled, memSize = NULL)
}
\arguments{
\item{enabled}{a logical}

\item{memSize}{to set the JVM size in Mb if it is not instantiated yet. This is optional. If not specified, the JVM is instantiated with
the default size.}
}
\description{
The multi threading allows to send multiple requests simultaneously. It
is handled by the back-end Java client. It is enabled by default.
}
