% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientConfiguration.R
\docType{data}
\name{settingEnv}
\alias{settingEnv}
\title{The settings environment for this package}
\format{
An object of class \code{environment} of length 0.
}
\usage{
settingEnv
}
\description{
This environment contains the general settings of the package.
}
\keyword{datasets}
