% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientConfiguration.R
\name{shutdownJava}
\alias{shutdownJava}
\title{Shut down the Java server}
\usage{
shutdownJava()
}
\description{
This method overrides the original function of the J4R package. It only adds
a call to the clearCache function before calling the original function of
the J4R package.
}
\examples{
\dontrun{
shutdownJava()}

}
