% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientFunctions.R
\docType{data}
\name{twoLocationsInSouthernQuebec}
\alias{twoLocationsInSouthernQuebec}
\title{A list of two plots located in southern Quebec}
\format{
An object of class \code{data.frame} with 2 rows and 5 columns.
}
\usage{
data(twoLocationsInSouthernQuebec)
}
\description{
A list of two plots located in southern Quebec
}
\examples{
data(twoLocationsInSouthernQuebec)
}
\keyword{datasets}
