% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientFunctions.R
\name{getModelList}
\alias{getModelList}
\title{Return the list of models available in BioSim}
\usage{
getModelList()
}
\description{
Provide the list of model that can be used in BioSIM after generating
the climate for some locations.
}
\examples{

\dontrun{
getModelList()}

}
