% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientFunctions.R
\name{getModelList}
\alias{getModelList}
\title{Returns the list of models available in BioSim for post weather generation processing.}
\usage{
getModelList(memSize = NULL)
}
\arguments{
\item{memSize}{the size of the Java Virtual Machine (optional)}
}
\description{
Returns the list of models available in BioSim for post weather generation processing.
}
