% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientFunctions.R
\name{getMonthlyNormals}
\alias{getMonthlyNormals}
\title{Return the monthly normals for a period}
\usage{
getMonthlyNormals(period, id, latDeg, longDeg, elevM = rep(NA,
  length(longDeg)), rcp = "RCP45", climModel = "RCM4")
}
\arguments{
\item{period}{a string representing the period (either "1951_1980", "1961_1990", "1971_2000", "1981_2010" up to "2071_2100")}

\item{id}{a vector with the ids of the plots}

\item{latDeg}{the latitudes of the plots}

\item{longDeg}{the longitudes of the plots}

\item{elevM}{the elevations of the plots (can contain some NA, in which case BioSim relies on a digital elevation model)}

\item{rcp}{an representative concentration pathway (either "RCP45" or "RCP85")}

\item{climModel}{a climatic model (either "RCM4", "GCM4" or "Hadley")}
}
\value{
a data.frame object
}
\description{
Return the monthly normals for a period
}
\examples{

locations <- BioSIM::twoLocationsInSouthernQuebec
\dontrun{
monthlyMeans <- getMonthlyNormals("1981_2010", locations$Name, locations$Latitude,
                                   locations$Longitude, locations$Elevation)}

}
