% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientFunctions.R
\name{isMultithreadingEnabled}
\alias{isMultithreadingEnabled}
\title{Checks whether or not the multi threading is enabled.}
\usage{
isMultithreadingEnabled(memSize = NULL)
}
\arguments{
\item{memSize}{to set the JVM size in Mb if it is not instantiated yet. This is optional. If not specified, the JVM is instantiated with
the default size.}
}
\value{
a logical
}
\description{
The multi threading allows to send multiple requests simultaneously. It
is handled by the back-end Java client. This method returns true if
this option is enabled (which is the default value).
}
