% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioSimClientConfiguration.R
\name{setMultithreadingEnabled}
\alias{setMultithreadingEnabled}
\title{Enables or disables the multi threading.}
\usage{
setMultithreadingEnabled(enabled)
}
\arguments{
\item{enabled}{a logical}
}
\description{
The multi threading allows to send multiple requests simultaneously. It
is handled by the back-end Java client. It is enabled by default.
}
